refactor(#798): refactor code and add new tests
