+++ deps/wazuh_testing/wazuh_testing/modules/vulnerability_detector/__init__.py
@@ -19,8 +19,8 @@ T_360 = 360
 T_800 = 800
 
 CUSTOM_VULNERABLE_PACKAGES = 'custom_vulnerable_packages.json'
-CUSTOM_VULNERABLE_PKG_MISSING_VENDOR = 'custom_vulnerable_pkg_missing_vendor.json'
-CUSTOM_VULNERABLE_PKG_MISSING_VENDOR_VERSION = 'custom_vulnerable_pkg_missing_vendor_version.json'
+CUSTOM_VULNERABLE_PKG_EMPTY_VENDOR = 'custom_vulnerable_pkg_empty_vendor.json'
+CUSTOM_VULNERABLE_PKG_EMPTY_VENDOR_VERSION = 'custom_vulnerable_pkg_empty_vendor_version.json'
 CUSTOM_NVD_FEED = 'custom_nvd_feed.json'
 CUSTOM_NVD_ALTERNATIVE_FEED = 'custom_nvd_alternative_feed.json'
 CUSTOM_REDHAT_JSON_FEED = 'custom_redhat_json_feed.json'
