+++ tests/integration/test_vulnerability_detector/test_cpe_helper/data/test_cases/cases_cpe_indexing_empty_vendor_version.yaml
@@ -1,4 +1,4 @@
-- name: Indexing CPE helper with missing all the source fields and version translation field
+- name: Missing all the source fields and version translation field
   description: Indexing CPE helper with missing all the source fields and version translation field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
