+++ tests/integration/test_vulnerability_detector/test_cpe_helper/data/test_cases/cases_cpe_indexing_missing_fields.yaml
@@ -1,4 +1,4 @@
-- name: Indexing CPE helper with missing version field
+- name: Missing version field
   description: Indexing CPE helper with missing version field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -54,7 +54,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing format_version field
+- name: Missing format_version field
   description: Indexing CPE helper with missing format_version field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -110,7 +110,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing update_date field
+- name: Missing update_date field
   description: Indexing CPE helper with missing update_date field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -166,7 +166,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing target field
+- name: Missing target field
   description: Indexing CPE helper with missing target field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -222,7 +222,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing action field
+- name: Missing action field
   description: Indexing CPE helper with missing action field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -278,7 +278,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing vendor field
+- name: Missing vendor field
   description: Indexing CPE helper with missing vendor field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -334,7 +334,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with missing product field
+- name: Missing product field
   description: Indexing CPE helper with missing product field
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
