+++ tests/integration/test_vulnerability_detector/test_cpe_helper/data/test_cases/cases_cpe_indexing_wrong_values.yaml
@@ -1,4 +1,4 @@
-- name: Indexing CPE helper with wrong version value
+- name: Wrong version value
   description: Indexing CPE helper with wrong version value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -53,7 +53,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong update_date value
+- name: Wrong update_date value
   description: Indexing CPE helper with wrong update_date value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -108,7 +108,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong target value
+- name: Wrong target value
   description: Indexing CPE helper with wrong target value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -163,7 +163,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong format_version value
+- name: Wrong format_version value
   description: Indexing CPE helper with wrong format_version value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -218,7 +218,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong source vendor value
+- name: Wrong source vendor value
   description: Indexing CPE helper with wrong source vendor value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -273,7 +273,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong source product value
+- name: Wrong source product value
   description: Indexing CPE helper with wrong source product value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
@@ -328,7 +328,7 @@
       DATE_VALUE: March 6, 2019.
       TYPE_VALUE: GPLv2
 
-- name: Indexing CPE helper with wrong action value
+- name: Wrong action value
   description: Indexing CPE helper with wrong action value
   configuration_parameters:
     NVD_JSON_PATH: CUSTOM_NVD_JSON_PATH
