+++ tests/integration/test_vulnerability_detector/test_cpe_helper/test_cpe_helper.py
@@ -73,11 +73,19 @@ t2_configurations_path = os.path.join(CONFIGURATIONS_PATH, 'configuration_cpe_in
 t2_cases_path = os.path.join(TEST_CASES_PATH, 'cases_cpe_indexing_wrong_values.yaml')
 t3_configurations_path = os.path.join(CONFIGURATIONS_PATH, 'configuration_cpe_indexing.yaml')
 t3_cases_path = os.path.join(TEST_CASES_PATH, 'cases_cpe_indexing_missing_fields.yaml')
+t4_configurations_path = os.path.join(CONFIGURATIONS_PATH, 'configuration_cpe_indexing.yaml')
+t4_cases_path = os.path.join(TEST_CASES_PATH, 'cases_cpe_indexing_empty_fields.yaml')
+t5_configurations_path = os.path.join(CONFIGURATIONS_PATH, 'configuration_cpe_indexing.yaml')
+t5_cases_path = os.path.join(TEST_CASES_PATH, 'cases_cpe_indexing_empty_vendor_version.yaml')
 
 # Custom paths
 custom_nvd_json_feed_path = os.path.join(TEST_FEEDS_PATH, 'nvd', vd.CUSTOM_NVD_FEED)
 custom_cpe_helper_path = os.path.join(TEST_FEEDS_PATH, 'cpe_helper', vd.CUSTOM_CPE_HELPER_TEMPLATE)
 custom_vulnerable_packages_path = os.path.join(TEST_PACKAGES_PATH, vd.CUSTOM_VULNERABLE_PACKAGES)
+custom_vulnerable_pkg_empty_vendor_path = os.path.join(TEST_PACKAGES_PATH,
+                                                       vd.CUSTOM_VULNERABLE_PKG_EMPTY_VENDOR)
+custom_vulnerable_pkg_empty_vendor_version_path = os.path.join(TEST_PACKAGES_PATH,
+                                                               vd.CUSTOM_VULNERABLE_PKG_EMPTY_VENDOR_VERSION)
 
 # CPE indexing packages test configurations (t1)
 t1_configuration_parameters, t1_configuration_metadata, t1_test_case_ids = get_test_cases_data(t1_cases_path)
@@ -97,6 +105,18 @@ t3_configurations = load_configuration_template(t3_configurations_path, t3_confi
                                                 t3_configuration_metadata)
 t3_systems = [metadata['system'] for metadata in t3_configuration_metadata]
 
+# CPE indexing packages test configurations (t4)
+t4_configuration_parameters, t4_configuration_metadata, t4_test_case_ids = get_test_cases_data(t4_cases_path)
+t4_configurations = load_configuration_template(t4_configurations_path, t4_configuration_parameters,
+                                                t4_configuration_metadata)
+t4_systems = [metadata['system'] for metadata in t4_configuration_metadata]
+
+# CPE indexing packages test configurations (t5)
+t5_configuration_parameters, t5_configuration_metadata, t5_test_case_ids = get_test_cases_data(t5_cases_path)
+t5_configurations = load_configuration_template(t5_configurations_path, t5_configuration_parameters,
+                                                t5_configuration_metadata)
+t5_systems = [metadata['system'] for metadata in t5_configuration_metadata]
+
 # Set offline custom feeds configuration
 t1_configurations = update_configuration_template(t1_configurations, ['CUSTOM_NVD_JSON_PATH'],
                                                   [custom_nvd_json_feed_path])
@@ -104,10 +124,19 @@ t2_configurations = update_configuration_template(t2_configurations, ['CUSTOM_NV
                                                   [custom_nvd_json_feed_path])
 t3_configurations = update_configuration_template(t3_configurations, ['CUSTOM_NVD_JSON_PATH'],
                                                   [custom_nvd_json_feed_path])
+t4_configurations = update_configuration_template(t4_configurations, ['CUSTOM_NVD_JSON_PATH'],
+                                                  [custom_nvd_json_feed_path])
+t5_configurations = update_configuration_template(t5_configurations, ['CUSTOM_NVD_JSON_PATH'],
+                                                  [custom_nvd_json_feed_path])
 
 # Global vars
-agent_packages = read_json_file(custom_vulnerable_packages_path)
-
+t1_agent_packages = [read_json_file(custom_vulnerable_packages_path) for metadata in t1_configuration_metadata]
+t2_agent_packages = [read_json_file(custom_vulnerable_packages_path) for metadata in t2_configuration_metadata]
+t3_agent_packages = [read_json_file(custom_vulnerable_packages_path) for metadata in t3_configuration_metadata]
+t4_agent_packages = [read_json_file(custom_vulnerable_pkg_empty_vendor_path)
+                     for metadata in t4_configuration_metadata]
+t5_agent_packages = [read_json_file(custom_vulnerable_pkg_empty_vendor_version_path)
+                     for metadata in t5_configuration_metadata]
 
 def replace_cpe_json_fields(tags=None, values=None):
     """Replace the tags and values of the generic_custom_cpe_helper.json file.
@@ -169,22 +198,37 @@ def remove_cpe_json_fields(tags=None):
 
 
 @pytest.fixture(scope='function')
-def prepare_environment(request, metadata, agent_system, mock_agent_with_custom_system):
+def prepare_environment(request, metadata, agent_system, agent_packages, mock_agent_with_custom_system):
     """Prepare the environment with a mocked agent, vulnerable packages and a custom cpe_helper.
 
     - Mock an agent with a specified system.
     - Insert mocked vulnerables packages.
     - Update packages sync status.
     - Copy the custom CPE helper to the dictionaries folder.
+    - Force full scan.
 
     Args:
         metadata (dict): Test case metadata.
         agent_system (str): System to set to the mocked agent.
+        agent_packages (list): List of vulnerable packages
         mock_agent_with_custom_system (fixture): Mock an agent with a custom system.
     """
     for package in agent_packages:
-        agent_db.insert_package(name=package['name'], version=package['version'], source=package['name'],
-                                vendor=package['vendor'], agent_id=mock_agent_with_custom_system)
+        try:
+            version = package['version']
+        except KeyError:
+            version = ''
+        try:
+            format = package['format']
+        except KeyError:
+            format = 'rpm'
+        try:
+            architecture = package['architecture']
+        except KeyError:
+            architecture = 'x64'
+        agent_db.insert_package(name=package['name'], format=format, architecture=architecture,
+                                agent_id=mock_agent_with_custom_system, vendor=package['vendor'],
+                                version=version, source=package['name'])
 
     # Sync packages info
     agent_db.update_sync_info(agent_id=mock_agent_with_custom_system, component="syscollector-packages")
@@ -208,9 +252,10 @@ def prepare_environment(request, metadata, agent_system, mock_agent_with_custom_
     write_json_file(CPE_HELPER_PATH, cpe_helper_backup_data)
 
 
-@pytest.mark.parametrize('configuration, metadata, agent_system',
-                         zip(t1_configurations, t1_configuration_metadata, t1_systems), ids=t1_test_case_ids)
-def test_cpe_indexing_wrong_tags(configuration, metadata, agent_system, set_wazuh_configuration_vdt,
+@pytest.mark.parametrize('configuration, metadata, agent_system, agent_packages',
+                         zip(t1_configurations, t1_configuration_metadata, t1_systems, t1_agent_packages),
+                         ids=t1_test_case_ids)
+def test_cpe_indexing_wrong_tags(configuration, metadata, agent_system, agent_packages, set_wazuh_configuration_vdt,
                                  truncate_monitored_files, clean_cve_tables_func, prepare_environment,
                                  restart_modulesd_function):
     '''
@@ -232,7 +277,7 @@ def test_cpe_indexing_wrong_tags(configuration, metadata, agent_system, set_wazu
         - Restore initial configuration, both ossec.conf and local_internal_options.conf.
         - Restore the original cpe_helper.json
 
-    wazuh_min_version: 4.4.0
+    wazuh_min_version: 4.5.0
 
     tier: 2
 
@@ -246,6 +291,9 @@ def test_cpe_indexing_wrong_tags(configuration, metadata, agent_system, set_wazu
         - agent_system:
             type: str
             brief: System to set to the mocked agent.
+        - agent_packages
+            type: list
+            brief: List of vulnerable packages.
         - set_wazuh_configuration_vdt:
             type: fixture
             brief: Set the wazuh configuration according to the configuration data.
@@ -293,9 +341,10 @@ def test_cpe_indexing_wrong_tags(configuration, metadata, agent_system, set_wazu
                 raise AttributeError('Unexpected log')
 
 
-@pytest.mark.parametrize('configuration, metadata, agent_system',
-                         zip(t2_configurations, t2_configuration_metadata, t2_systems), ids=t2_test_case_ids)
-def test_cpe_indexing_wrong_values(configuration, metadata, agent_system, set_wazuh_configuration_vdt,
+@pytest.mark.parametrize('configuration, metadata, agent_system, agent_packages',
+                         zip(t2_configurations, t2_configuration_metadata, t2_systems, t2_agent_packages),
+                         ids=t2_test_case_ids)
+def test_cpe_indexing_wrong_values(configuration, metadata, agent_system, agent_packages, set_wazuh_configuration_vdt,
                                    truncate_monitored_files, clean_cve_tables_func, prepare_environment,
                                    restart_modulesd_function):
     '''
@@ -317,7 +366,7 @@ def test_cpe_indexing_wrong_values(configuration, metadata, agent_system, set_wa
         - Restore initial configuration, both ossec.conf and local_internal_options.conf.
         - Restore the original cpe_helper.json
 
-    wazuh_min_version: 4.4.0
+    wazuh_min_version: 4.5.0
 
     tier: 2
 
@@ -331,6 +380,9 @@ def test_cpe_indexing_wrong_values(configuration, metadata, agent_system, set_wa
         - agent_system:
             type: str
             brief: System to set to the mocked agent.
+        - agent_packages
+            type: list
+            brief: List of vulnerable packages.
         - set_wazuh_configuration_vdt:
             type: fixture
             brief: Set the wazuh configuration according to the configuration data.
@@ -378,9 +430,10 @@ def test_cpe_indexing_wrong_values(configuration, metadata, agent_system, set_wa
                 raise AttributeError('Unexpected log')
 
 
-@pytest.mark.parametrize('configuration, metadata, agent_system',
-                         zip(t3_configurations, t3_configuration_metadata, t3_systems), ids=t3_test_case_ids)
-def test_cpe_indexing_missing_field(configuration, metadata, agent_system, set_wazuh_configuration_vdt,
+@pytest.mark.parametrize('configuration, metadata, agent_system, agent_packages',
+                         zip(t3_configurations, t3_configuration_metadata, t3_systems, t3_agent_packages),
+                         ids=t3_test_case_ids)
+def test_cpe_indexing_missing_field(configuration, metadata, agent_system, agent_packages, set_wazuh_configuration_vdt,
                                     truncate_monitored_files, clean_cve_tables_func, prepare_environment,
                                     restart_modulesd_function):
     '''
@@ -402,7 +455,7 @@ def test_cpe_indexing_missing_field(configuration, metadata, agent_system, set_w
         - Restore initial configuration, both ossec.conf and local_internal_options.conf.
         - Restore the original cpe_helper.json
 
-    wazuh_min_version: 4.4.0
+    wazuh_min_version: 4.5.0
 
     tier: 2
 
@@ -416,6 +469,9 @@ def test_cpe_indexing_missing_field(configuration, metadata, agent_system, set_w
         - agent_system:
             type: str
             brief: System to set to the mocked agent.
+        - agent_packages
+            type: list
+            brief: List of vulnerable packages.
         - set_wazuh_configuration_vdt:
             type: fixture
             brief: Set the wazuh configuration according to the configuration data.
@@ -460,3 +516,154 @@ def test_cpe_indexing_missing_field(configuration, metadata, agent_system, set_w
                 raise AttributeError('Unexpected log')
     elif expected_result == 'error_inserting_package':
         evm.check_error_inserting_package(agent_id=prepare_environment)
+
+
+@pytest.mark.parametrize('configuration, metadata, agent_system, agent_packages',
+                         zip(t4_configurations, t4_configuration_metadata, t4_systems, t4_agent_packages),
+                         ids=t4_test_case_ids)
+def test_cpe_indexing_empty_fields(configuration, metadata, agent_system, agent_packages, set_wazuh_configuration_vdt,
+                                   truncate_monitored_files, clean_cve_tables_func, prepare_environment,
+                                   restart_modulesd_function):
+    '''
+    description: Check if the packages are indexed in the database by checking the respective log in the ossec.log file,
+                 and if the alert of the vulnerable package comes out when some tag are empty.
+
+    test_phases:
+    - setup:
+        - Load Wazuh light configuration, with custom feeds.
+        - Apply ossec.conf configuration changes according to the configuration template and use case.
+        - Apply custom settings in local_internal_options.conf.
+        - Mock an agent with Windows system and vulnerable packages.
+        - Backup the original cpe_helper.json and copy a custom CPE helper with new tags and values.
+        - Truncate wazuh logs.
+        - Restart wazuh-modulesd daemon to apply configuration changes.
+    - test:
+        - Check in the log and alert for specific information.
+    - teardown:
+        - Truncate wazuh logs.
+        - Restore initial configuration, both ossec.conf and local_internal_options.conf.
+        - Restore the original cpe_helper.json
+
+    wazuh_min_version: 4.5.0
+
+    tier: 1
+
+    parameters:
+        - configuration:
+            type: dict
+            brief: Wazuh configuration data. Needed for set_wazuh_configuration fixture.
+        - metadata:
+            type: dict
+            brief: Wazuh configuration metadata.
+        - agent_system:
+            type: str
+            brief: System to set to the mocked agent.
+        - agent_packages
+            type: list
+            brief: List of vulnerable packages.
+        - set_wazuh_configuration_vdt:
+            type: fixture
+            brief: Set the wazuh configuration according to the configuration data.
+        - truncate_monitored_files:
+            type: fixture
+            brief: Truncate all the log files and json alerts files before and after the test execution.
+        - clean_cve_tables_func:
+            type: fixture
+            brief: Clean all CVE tables.
+        - prepare_environment:
+            type: fixture
+            brief: Setup the initial test state.
+        - restart_modulesd_function:
+            type: fixture
+            brief: Restart the wazuh-modulesd daemon.
+
+    assertions:
+        - Check for a specific log and alert.
+
+    input_description:
+        - The `configuration_cpe_indexing.yaml` file provides the module configuration for this test.
+        - The `cases_cpe_indexing_missing_fields.yaml` file provides the test cases.
+
+    expected_output:
+        - r"The CPE .*a:{package_vendor}:{package_name}.* from the agent '{agent_id}' was indexed"
+        - fr".*"agent":."id":"{agent_id}".*{cve} affects {package}', prefix='.*"
+    '''
+    for package in agent_packages:
+        evm.check_cpe_helper_packages_indexed(package_name=metadata['values']['PRODUCT_T_VALUE_0'],
+                                              package_vendor=metadata['values']['VENDOR_T_VALUE'],
+                                              agent_id=prepare_environment, timeout=vd.T_20)
+
+        evm.check_vulnerability_affects_alert(agent_id=prepare_environment,
+                                              package=metadata['values']['PRODUCT_T_VALUE_0'], cve=package['cveid'])
+
+
+@pytest.mark.parametrize('configuration, metadata, agent_system, agent_packages',
+                         zip(t5_configurations, t5_configuration_metadata, t5_systems, t5_agent_packages),
+                         ids=t5_test_case_ids)
+def test_cpe_indexing_empty_vendor_version(configuration, metadata, agent_system, agent_packages,
+                                           set_wazuh_configuration_vdt, truncate_monitored_files,
+                                           clean_cve_tables_func, prepare_environment, restart_modulesd_function):
+    '''
+    description: Check that when vendor and version tags are empty, and the action tag is not the correct to
+                 extract the version field, the package cannot be indexed.
+
+    test_phases:
+    - setup:
+        - Load Wazuh light configuration, with custom feeds.
+        - Apply ossec.conf configuration changes according to the configuration template and use case.
+        - Apply custom settings in local_internal_options.conf.
+        - Mock an agent with Windows system and vulnerable packages.
+        - Backup the original cpe_helper.json and copy a custom CPE helper with new tags and values.
+        - Truncate wazuh logs.
+        - Restart wazuh-modulesd daemon to apply configuration changes.
+    - test:
+        - Check in the log and alert for specific information.
+    - teardown:
+        - Truncate wazuh logs.
+        - Restore initial configuration, both ossec.conf and local_internal_options.conf.
+        - Restore the original cpe_helper.json
+
+    wazuh_min_version: 4.5.0
+
+    tier: 1
+
+    parameters:
+        - configuration:
+            type: dict
+            brief: Wazuh configuration data. Needed for set_wazuh_configuration fixture.
+        - metadata:
+            type: dict
+            brief: Wazuh configuration metadata.
+        - agent_system:
+            type: str
+            brief: System to set to the mocked agent.
+        - agent_packages
+            type: list
+            brief: List of vulnerable packages.
+        - set_wazuh_configuration_vdt:
+            type: fixture
+            brief: Set the wazuh configuration according to the configuration data.
+        - truncate_monitored_files:
+            type: fixture
+            brief: Truncate all the log files and json alerts files before and after the test execution.
+        - clean_cve_tables_func:
+            type: fixture
+            brief: Clean all CVE tables.
+        - prepare_environment:
+            type: fixture
+            brief: Setup the initial test state.
+        - restart_modulesd_function:
+            type: fixture
+            brief: Restart the wazuh-modulesd daemon.
+
+    assertions:
+        - Check for a specific log and alert.
+
+    input_description:
+        - The `configuration_cpe_indexing.yaml` file provides the module configuration for this test.
+        - The `cases_cpe_indexing_missing_vendor_version.yaml` file provides the test cases.
+
+    expected_output:
+        - fr"DEBUG: .* Couldn't get the version of the CPE for the {package_name} package."
+    '''
+    evm.check_version_log(package_name=metadata['values']['PRODUCT_T_VALUE_0'])
